Refactor unit tests to include the new ingest service
