--- tests/unit_tests/routers/ingest/test_collections.json
@@ -1,14 +1,0 @@
-{
-  "collections": [
-    {
-      "import_id": "test.new.collection.0",
-      "title": "Test title",
-      "description": "Test description"
-    },
-    {
-      "import_id": "test.new.collection.1",
-      "title": "Test title",
-      "description": "Test description"
-    }
-  ]
-}
